refactor: LottoGenerator 역할을 Lotto와 Lottos로 분리
