+++ src/main/java/lotto/model/Lotto.java
@@ -2,8 +2,14 @@ package lotto.model;
 
 import java.util.Collections;
 import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
 
 public class Lotto {
+    private static final int MIN_LOTTO_NUMBER = 1;
+    private static final int MAX_LOTTO_NUMBER = 45;
+    private static final int LOTTO_SIZE = 6;
+
     private final List<Integer> lottoNumbers;
 
     public Lotto(List<Integer> lottoNumbers) {
@@ -23,6 +29,19 @@ public class Lotto {
 
     public boolean matchBonusNumber(Integer bonusNumber) {
         return lottoNumbers.contains(bonusNumber);
+    }
+
+    public static Lotto generate() {
+        List<Integer> lottoNumbers = generateSequentialIntegers();
+        Collections.shuffle(lottoNumbers);
+
+        return new Lotto(lottoNumbers.subList(0, LOTTO_SIZE));
+    }
+
+    private static List<Integer> generateSequentialIntegers() {
+        return IntStream.rangeClosed(MIN_LOTTO_NUMBER, MAX_LOTTO_NUMBER)
+                .boxed()
+                .collect(Collectors.toList());
     }
 
     public List<Integer> getLottoNumbers() {
