+++ src/main/java/lotto/model/LottoGame.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public class LottoGame {
     private static final int LOTTO_PRICE = 1000;
-    public static final String ERROR_NOT_DIVIDED_BY_UNIT_PRICE = "거스름돈을 지급하지 않습니다. 금액이 남지 않게 지불해주세요.";
+    private static final String ERROR_NOT_DIVIDED_BY_UNIT_PRICE = "거스름돈을 지급하지 않습니다. 금액이 남지 않게 지불해주세요.";
 
     private final LottoMatcher lottoMatcher;
 
@@ -37,9 +37,8 @@ public class LottoGame {
 
     public static Lottos buyLottos(int money) {
         int lottoSize = getLottoSize(money);
-        LottoGenerator lottoGenerator = new LottoGenerator();
 
-        return lottoGenerator.generateLottos(lottoSize);
+        return Lottos.generate(lottoSize);
     }
 
     public ResultMap getWinningResult(Lottos lottos) {
