+++ src/main/java/lotto/model/WinningNumbers.java
@@ -8,6 +8,9 @@ public class WinningNumbers {
     private static final String ERROR_NOT_MATCH_WINNING_NUMBER_SIZE = "지난 주 당첨 번호 개수는 6개로 입력해주세요.";
     private static final String ERROR_OUT_OF_RANGE_WINNING_NUMBERS = "당첨 번호의 범위는 1 ~ 45 사이로 입력해주세요.";
     private static final String ERROR_DUPLICATION_WINNING_NUMBERS = "당첨 번호에 중복이 존재합니다.";
+    private static final int MIN_LOTTO_NUMBER = 1;
+    private static final int MAX_LOTTO_NUMBER = 45;
+    private static final int LOTTO_SIZE = 6;
 
     private final List<Integer> winningNumbers;
 
@@ -19,14 +22,14 @@ public class WinningNumbers {
     }
 
     private void validateNumberOfWinningNumbers(List<Integer> winningNumbers) {
-        if (winningNumbers.size() != 6) {
+        if (winningNumbers.size() != LOTTO_SIZE) {
             throw new IllegalArgumentException(ERROR_NOT_MATCH_WINNING_NUMBER_SIZE);
         }
     }
 
     private void validateRangeWinningNumbers(List<Integer> winningNumbers) {
         winningNumbers.forEach(number -> {
-            if (number < 1 || number > 45) {
+            if (number < MIN_LOTTO_NUMBER || number > MAX_LOTTO_NUMBER) {
                 throw new IllegalArgumentException(ERROR_OUT_OF_RANGE_WINNING_NUMBERS);
             }
         });
