--- src/test/java/lotto/model/LottoGeneratorTest.java
@@ -1,32 +1,0 @@
-package lotto.model;
-
-import lotto.model.Lotto;
-import lotto.model.LottoGenerator;
-import lotto.model.Lottos;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-import java.util.List;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-class LottoGeneratorTest {
-
-    LottoGenerator lottoGenerator = new LottoGenerator();
-
-    @Test
-    @DisplayName("6개의 랜덤한 로또 번호를 생성하는지 확인")
-    void shuffleTest() {
-        Lotto lotto = lottoGenerator.generateLotto();
-
-        assertThat(lotto.getSize()).isEqualTo(6);
-    }
-
-    @Test
-    @DisplayName("3개의 로또 생성하는지 확인")
-    void GenerateLottos() {
-        Lottos lottos = lottoGenerator.generateLottos(3);
-
-        assertThat(lottos.getLottos().size()).isEqualTo(3);
-    }
-}
